# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Pet Platform Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge/breeds": {
      "get": {
        "operationId": "public_breeds_api_v1_knowledge_breeds_get",
        "parameters": [
          {
            "in": "query",
            "name": "species",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(dog|cat)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Species"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Public Breeds Api V1 Knowledge Breeds Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Public Breeds",
        "tags": [
          "pet-knowledge"
        ]
      }
    },
    "/api/v1/knowledge/breeds/{breed_id}": {
      "get": {
        "operationId": "public_breed_detail_api_v1_knowledge_breeds__breed_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "breed_id",
            "required": true,
            "schema": {
              "title": "Breed Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Public Breed Detail Api V1 Knowledge Breeds  Breed Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Public Breed Detail",
        "tags": [
          "pet-knowledge"
        ]
      }
    },
    "/api/v1/knowledge/pets/{pet_id}": {
      "get": {
        "operationId": "pet_knowledge_api_v1_knowledge_pets__pet_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pet Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Pet Knowledge Api V1 Knowledge Pets  Pet Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Pet Knowledge",
        "tags": [
          "pet-knowledge"
        ]
      }
    },
    "/api/v1/knowledge/search": {
      "get": {
        "operationId": "search_breeds_api_v1_knowledge_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "species",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(dog|cat)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Species"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Search Breeds Api V1 Knowledge Search Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Breeds",
        "tags": [
          "pet-knowledge"
        ]
      }
    },
...
  }
}